+++ NEWS.md
@@ -1,3 +1,5 @@
+* Adjusted weight calculated automatically in the shiny app
+
 # equationer (development version)
 
 * Hot-fix a sign erorr in a eqaution (#17)